Rescan download folder after copying over files
